Last HW
I spent the better part of friday and all weekend trying to figure out
how to do number 2. But I don’t get it. I understand conceptually that
it needs to be assigned to a pool and then that pool can be split up to
run in parallel. But I cannot for the life of me figure out how to make
that work.